+++ src/voronoidiagram/index.js
@@ -1,0 +1,1 @@
+export { default } from './voronoidiagram' // update the name of the folder