+++ src/voronoidiagram/mapping.js
@@ -1,0 +1,6 @@
+export const mapData = {
+  x: 'get',
+  y: 'get',
+  color: 'get',
+  label: 'get',
+}