+++ src/voronoidiagram/metadata.js
@@ -1,0 +1,14 @@
+import icon from './voronoidiagram.svg'
+import thumbnail from './voronoidiagram_thumb.svg'
+
+export const metadata = {
+  name: 'Voronoi Diagram',
+  id: 'rawgraphs.voronoidiagram',
+  thumbnail,
+  icon,
+  categories: ['Dispersions'],
+  description:
+    'It creates the minimum area around each point defined by two variables. When applied to a scatterplot, it is useful to show the distance between points.',
+  // code: 'https://github.com/rawgraphs',
+  // tutorial: 'https://rawgraphs.io/learning/',
+}