+++ src/contourPlot/visualOptions.js
@@ -23,7 +23,7 @@ export const visualOptions = {
   marginLeft: {
     type: 'number',
     label: 'Margin (left)',
-    default: 100,
+    default: 50,
     group: 'artboard',
   },